// src/state/types.ts
export interface NFTVODAReferral {
  id: string
  claimed: boolean
  reward: string
  time: number
}